Update vpn script to allow keeping vpn up
Sometimes vpn container failed (network blip) and not comes back as it uses
OTP. This change allow them to reset itself and keep vpn up all the time.

The run in /etc/rc.local seems not working after system reboot but it works
when run as ssh root console. Not sure why yet.