Update vpn script
This allow us to run as a cron task for 5 minutes period and re-start the
script if it exited.

This is to fix the wa vpn as pritunl there is only up on bisness hours. Thus
the script will exited at some stage and then restarted.

As run as cron job and remove form rc.local - thus at boot after 5 minutes at
most these vpn will automatically started. We no longer need to ssh in and run it manually.